Fork tdfsb to 3dfsb
I've tried getting in touch with the original author for quite some time via different channels, but I haven't received any reply set.

To be able to continue, I've decided to fork and rename the project to make a distinction, considering that many things have changed, such as the dependencies and the (as of now) untested compatibility with BeOS and Windows.