Open the URL field of .desktop files that have it with xdg-open
To make a shortcut to an URL, users often make a .desktop file with a URL= field.
These .desktop files are just text files, so xdg-open used to open them with a text editor.
This patch searches for a URL= field in .desktop files and if found, does xdg-open of that URL, which typically starts the default browser.